Move owner test helper methods to util and factor out common code
